cast off_t to long long int for printing
reduces the diff with opensmtpd-portable.
ok tb@

+++ libexec/mail.local/mail.local.c
@@ -1,4 +1,4 @@
-/*	$OpenBSD: mail.local.c,v 1.38 2020/02/07 02:27:37 millert Exp $	*/
+/*	$OpenBSD: mail.local.c,v 1.39 2020/02/09 14:59:20 millert Exp $	*/
 
 /*-
  * Copyright (c) 1996-1998 Theo de Raadt <deraadt@theos.com>
@@ -243,7 +243,8 @@ retry:
 	}
 
 	curoff = lseek(mbfd, 0, SEEK_END);
-	(void)snprintf(biffmsg, sizeof biffmsg, "%s@%lld\n", name, curoff);
+	(void)snprintf(biffmsg, sizeof biffmsg, "%s@%lld\n", name,
+	    (long long int)curoff);
 	if (lseek(fd, 0, SEEK_SET) == (off_t)-1) {
 		mwarn("temporary file: %s", strerror(errno));
 		goto bad;
